abstracted out ballLift.java
